% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{availableMotifSets}
\alias{availableMotifSets}
\title{List the motif sets provided by this package}
\usage{
availableMotifSets()
}
\value{
A character vector of motif set names.
}
\description{
Returns the names accepted by the \code{motifSet}
argument of \code{\link{getTFbindsites}} and
\code{\link{getGCfreq}}.
}
\examples{
availableMotifSets()
}
\seealso{
\code{\link{methylTFRAnnotationHg38}} for an overview of the
package.
}
