% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTFbindsites.R
\name{getTFbindsites}
\alias{getTFbindsites}
\title{getTFbindsites}
\usage{
getTFbindsites(motifSet = "altius")
}
\arguments{
\item{motifSet}{Motif set to load. One of: altius, cisbpv2, jaspar2020,
jaspar2020_distal.}
}
\value{
A \code{GRangesList}, one element per motif.
}
\description{
Retrieve transcription factor binding sites for a motif set.
Downloaded from AnnotationHub on first use and cached locally thereafter.
}
\examples{
# 1. Create a dummy object and save it to a temp directory
mock_dir <- tempdir()
mock_file <- file.path(mock_dir, "altius_tf_bindsites.rds")

# Mocking a GRangesList as expected by the function output
mock_grl <- GenomicRanges::GRangesList(
    mock_motif_1 = GenomicRanges::GRanges("chr1:100-200")
)
saveRDS(mock_grl, mock_file)

# 2. Point the local directory option to the temp directory
old_opt <- getOption("methylTFRAnnotationHg38.datadir")
options(methylTFRAnnotationHg38.datadir = mock_dir)

# 3. Run the function
x <- getTFbindsites()

# 4. Restore options and clean up
options(methylTFRAnnotationHg38.datadir = old_opt)
file.remove(mock_file)

}
